[Feat] Wave 준비 상태 시 Creature 처리 및 UI 연동

+++ Assets/Scripts/Managers/GameManager.cs
@@ -17,7 +17,7 @@ public class GameManager : MonoBehaviour
     Player _player;
     public Player Player { get { return _player; } }
 
-    [SerializeField] float readyTime;
+    [SerializeField] int readyTime;
 
     void Awake()
     {
